Fix bug when using multiline strings inside a block.
read_block used to stop reading when it found a triple-quote literal,
even if there was still stuff to read to that block. For example, when
entering:

    >>> def f():
    ...     a = """
    ... foo"""

You'd be presented with a normal prompt and be unable to continue
writing the rest of the function. This commit fixes that bug.